ci: sign and publish releases from GitHub Releases
Wires the updater end to end. tauri.conf.json now carries the public key
and points at releases/latest/download/latest.json, and
createUpdaterArtifacts emits the manifest the feed serves. The private
key stays outside the repo; only its public half is committed.

Releases are tag-driven and published as drafts, so an ordinary commit
can never advertise itself as an update and a release is readable before
every installed copy is offered it.

TAURI_SIGNING_PRIVATE_KEY_PASSWORD is set explicitly to an empty string,
which is not the same as leaving it unset. The key has no password, but
Tauri decrypts it on every signed build and prompts when the variable is
absent — on a runner that hangs the job at a prompt nobody can answer,
and the last line of output is "expect a prompt for password", which
reads as a slow build rather than a stuck one. This was found the hard
way; docs/UPDATES.md now names it.

Also adds a CI workflow for ordinary pushes: typecheck, both test suites,
a frontend build, and clippy. It deliberately does not bundle, because
bundling needs the signing secret and a check requiring a secret cannot
run on a fork's pull request.

Clippy is gated at -D warnings, which meant fixing the six warnings the
codebase had accumulated — it had never been run here.

Two tests guard the configuration itself: the version in tauri.conf.json
and APP_VERSION must agree, since drift makes the app either offer an
update to the version it already is or report itself current when it is
not, both of which look like a broken updater rather than a missed edit.
The other asserts no private key has been committed.

+++ src-tauri/src/commands/db.rs
@@ -5,9 +5,9 @@
 //!   - SQLite     — rusqlite (bundled), runs on the blocking pool.
 //!   - SQL Server — tiberius over tokio TCP (values extracted via a typed cascade).
 //!   - MySQL      — mysql_async.
-//!   - Oracle     — feature-gated (`--features oracle`); requires Oracle Instant Client at
-//!                  build + runtime, so it is OFF by default and the default build never
-//!                  links it. UI marks it accordingly.
+//!   - Oracle     — feature-gated (`--features oracle`); requires Oracle Instant Client
+//!     at build + runtime, so it is OFF by default and the default build never links it.
+//!     UI marks it accordingly.
 //!
 //! This is a read/execute surface for a developer tool: values are returned as strings for
 //! display. Destructive-statement gating and safe-mode live in the frontend.
@@ -453,7 +453,7 @@ fn mysql_value_to_string(v: &mysql_async::Value) -> Option<String> {
         Value::Time(neg, days, h, mi, s, us) => Some(format!(
             "{}{:02}:{:02}:{:02}.{:06}",
             if *neg { "-" } else { "" },
-            (*days as u32) * 24 + *h as u32,
+            *days * 24 + *h as u32,
             mi,
             s,
             us
